requirements.txt
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-vector-stores-chroma
chromadb
gradio
pyyaml
requests
python-dateutil
